Modified sieve function
it should be able to take in a parameter 'limit' as the input, then
output all prime numbers under this limit.
Because list is '0' starting, so the list has +1 in length, where the
index number is the number we are looking at. Therefore index '2' ->
number 2.
I use xrange instead of range because it's python 2, which by consensus
it is bad practice to use range...